basic calculation of required tiles, taking into account scaling and offsets
